refactored sqlite to better separate tests
